Refactor SearchResults component
